Ptorotype and externalize sidebar items for the sidebarmenu , minor change in components' descriptions
