+++ src/components/SidebarComponent/SidebarComponent.tsx
@@ -1,5 +1,4 @@
 import { styled, useTheme, Theme, CSSObject } from '@mui/material/styles';
-import { Link } from 'react-router-dom';
 import MuiDrawer from '@mui/material/Drawer';
 import List from '@mui/material/List';
 import Divider from '@mui/material/Divider';
@@ -7,14 +6,13 @@ import IconButton from '@mui/material/IconButton';
 import ChevronLeftIcon from '@mui/icons-material/ChevronLeft';
 import ChevronRightIcon from '@mui/icons-material/ChevronRight';
 import ListItem from '@mui/material/ListItem';
-import ListItemButton from '@mui/material/ListItemButton';
-import ListItemIcon from '@mui/material/ListItemIcon';
-import ListItemText from '@mui/material/ListItemText';
+
 import MapIcon from '@mui/icons-material/Map';
 import InfoIcon from '@mui/icons-material/Info';
 import ConnectWithoutContactIcon from '@mui/icons-material/ConnectWithoutContact';
 import DeviceHubIcon from '@mui/icons-material/DeviceHub';
 import SupportAgentIcon from '@mui/icons-material/SupportAgent';
+import SideBarItem from './SideBarItemComponent';
 
 // TODO: Export this globally
 const drawerWidth = 270;
@@ -79,11 +77,6 @@ const Drawer = styled(MuiDrawer, {
   }),
 }));
 
-const LinkWithText = styled(Link)(() => ({
-  color: 'inherit',
-  textDecoration: 'none',
-}));
-
 /**
  * Sidebar component
  * @param {boolean} open  - A flag representing whether or not the sidebar is open.
@@ -110,71 +103,28 @@ export default function SidebarComponent({
       <Divider />
       <List>
         {mainMenuItems.map((text, index) => (
-          // TODO: Make a reusable component here
           <ListItem key={text} disablePadding sx={{ display: 'block' }}>
             {index === 0 ? (
-              <LinkWithText to='/map'>
-                <ListItemButton
-                  sx={{
-                    minHeight: 48,
-                    justifyContent: open ? 'initial' : 'center',
-                    px: 2.5,
-                  }}
-                >
-                  <ListItemIcon
-                    sx={{
-                      minWidth: 0,
-                      mr: open ? 3 : 'auto',
-                      justifyContent: 'center',
-                    }}
-                  >
-                    <MapIcon />
-                  </ListItemIcon>
-                  <ListItemText primary={text} sx={{ opacity: open ? 1 : 0 }} />
-                </ListItemButton>
-              </LinkWithText>
+              <SideBarItem
+                open={open}
+                text={text}
+                url={'/map'}
+                icon={<MapIcon />}
+              />
             ) : index === 1 ? (
-              <LinkWithText to='/status'>
-                <ListItemButton
-                  sx={{
-                    minHeight: 48,
-                    justifyContent: open ? 'initial' : 'center',
-                    px: 2.5,
-                  }}
-                >
-                  <ListItemIcon
-                    sx={{
-                      minWidth: 0,
-                      mr: open ? 3 : 'auto',
-                      justifyContent: 'center',
-                    }}
-                  >
-                    <DeviceHubIcon />
-                  </ListItemIcon>
-                  <ListItemText primary={text} sx={{ opacity: open ? 1 : 0 }} />
-                </ListItemButton>
-              </LinkWithText>
+              <SideBarItem
+                open={open}
+                text={text}
+                url={'/status'}
+                icon={<DeviceHubIcon />}
+              />
             ) : (
-              <LinkWithText to='/proximity'>
-                <ListItemButton
-                  sx={{
-                    minHeight: 48,
-                    justifyContent: open ? 'initial' : 'center',
-                    px: 2.5,
-                  }}
-                >
-                  <ListItemIcon
-                    sx={{
-                      minWidth: 0,
-                      mr: open ? 3 : 'auto',
-                      justifyContent: 'center',
-                    }}
-                  >
-                    <ConnectWithoutContactIcon />
-                  </ListItemIcon>
-                  <ListItemText primary={text} sx={{ opacity: open ? 1 : 0 }} />
-                </ListItemButton>
-              </LinkWithText>
+              <SideBarItem
+                open={open}
+                text={text}
+                url={'/proximity'}
+                icon={<ConnectWithoutContactIcon />}
+              />
             )}
           </ListItem>
         ))}
@@ -184,47 +134,19 @@ export default function SidebarComponent({
         {secondaryMenuItems.map((text, index) => (
           <ListItem key={text} disablePadding sx={{ display: 'block' }}>
             {index === 0 ? (
-              <LinkWithText to='/contact'>
-                <ListItemButton
-                  sx={{
-                    minHeight: 48,
-                    justifyContent: open ? 'initial' : 'center',
-                    px: 2.5,
-                  }}
-                >
-                  <ListItemIcon
-                    sx={{
-                      minWidth: 0,
-                      mr: open ? 3 : 'auto',
-                      justifyContent: 'center',
-                    }}
-                  >
-                    <SupportAgentIcon />
-                  </ListItemIcon>
-                  <ListItemText primary={text} sx={{ opacity: open ? 1 : 0 }} />
-                </ListItemButton>
-              </LinkWithText>
+              <SideBarItem
+                open={open}
+                text={text}
+                url={'/contact'}
+                icon={<SupportAgentIcon />}
+              />
             ) : (
-              <LinkWithText to='/about'>
-                <ListItemButton
-                  sx={{
-                    minHeight: 48,
-                    justifyContent: open ? 'initial' : 'center',
-                    px: 2.5,
-                  }}
-                >
-                  <ListItemIcon
-                    sx={{
-                      minWidth: 0,
-                      mr: open ? 3 : 'auto',
-                      justifyContent: 'center',
-                    }}
-                  >
-                    <InfoIcon />
-                  </ListItemIcon>
-                  <ListItemText primary={text} sx={{ opacity: open ? 1 : 0 }} />
-                </ListItemButton>
-              </LinkWithText>
+              <SideBarItem
+                open={open}
+                text={text}
+                url={'/about'}
+                icon={<InfoIcon />}
+              />
             )}
           </ListItem>
         ))}
